refactor: remove UserManagementIndexDescriptor and move identity indices
