+++ search/search-client-query-transformer/src/main/java/io/camunda/search/clients/transformers/filter/UserFilterTransformer.java
@@ -11,10 +11,10 @@ import static io.camunda.search.clients.query.SearchQueryBuilders.and;
 import static io.camunda.search.clients.query.SearchQueryBuilders.matchNone;
 import static io.camunda.search.clients.query.SearchQueryBuilders.stringTerms;
 import static io.camunda.search.clients.query.SearchQueryBuilders.term;
-import static io.camunda.webapps.schema.descriptors.usermanagement.index.UserIndex.EMAIL;
-import static io.camunda.webapps.schema.descriptors.usermanagement.index.UserIndex.KEY;
-import static io.camunda.webapps.schema.descriptors.usermanagement.index.UserIndex.NAME;
-import static io.camunda.webapps.schema.descriptors.usermanagement.index.UserIndex.USERNAME;
+import static io.camunda.webapps.schema.descriptors.index.UserIndex.EMAIL;
+import static io.camunda.webapps.schema.descriptors.index.UserIndex.KEY;
+import static io.camunda.webapps.schema.descriptors.index.UserIndex.NAME;
+import static io.camunda.webapps.schema.descriptors.index.UserIndex.USERNAME;
 
 import io.camunda.search.clients.query.SearchQuery;
 import io.camunda.search.filter.UserFilter;
