+++ webapps-schema/src/main/java/io/camunda/webapps/schema/descriptors/index/GroupIndex.java
@@ -5,32 +5,28 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.webapps.schema.descriptors.usermanagement.index;
+package io.camunda.webapps.schema.descriptors.index;
 
+import io.camunda.webapps.schema.descriptors.AbstractIndexDescriptor;
+import io.camunda.webapps.schema.descriptors.ComponentNames;
 import io.camunda.webapps.schema.descriptors.backup.Prio5Backup;
-import io.camunda.webapps.schema.descriptors.usermanagement.UserManagementIndexDescriptor;
-import io.camunda.webapps.schema.entities.usermanagement.EntityJoinRelation;
 import io.camunda.webapps.schema.entities.usermanagement.EntityJoinRelation.EntityJoinRelationFactory;
+import io.camunda.webapps.schema.entities.usermanagement.EntityJoinRelation.IdentityJoinRelationshipType;
 
-public class TenantIndex extends UserManagementIndexDescriptor implements Prio5Backup {
+public class GroupIndex extends AbstractIndexDescriptor implements Prio5Backup {
 
-  public static final String INDEX_NAME = "tenant";
+  public static final String INDEX_NAME = "group";
   public static final String INDEX_VERSION = "8.8.0";
-
   public static final String KEY = "key";
-  public static final String TENANT_ID = "tenantId";
+  public static final String MEMBER_KEY = "memberKey";
   public static final String NAME = "name";
   public static final String JOIN = "join";
-  public static final String MEMBER_TYPE = "memberType";
-  public static final String MEMBER_ID = "memberId";
-  public static final String PARENT = "parent";
 
-  public static final EntityJoinRelationFactory<String> JOIN_RELATION_FACTORY =
+  public static final EntityJoinRelationFactory<Long> JOIN_RELATION_FACTORY =
       new EntityJoinRelationFactory<>(
-          EntityJoinRelation.IdentityJoinRelationshipType.TENANT,
-          EntityJoinRelation.IdentityJoinRelationshipType.MEMBER);
+          IdentityJoinRelationshipType.GROUP, IdentityJoinRelationshipType.MEMBER);
 
-  public TenantIndex(final String indexPrefix, final boolean isElasticsearch) {
+  public GroupIndex(final String indexPrefix, final boolean isElasticsearch) {
     super(indexPrefix, isElasticsearch);
   }
 
@@ -42,5 +38,10 @@ public class TenantIndex extends UserManagementIndexDescriptor implements Prio5B
   @Override
   public String getVersion() {
     return INDEX_VERSION;
+  }
+
+  @Override
+  public String getComponentName() {
+    return ComponentNames.CAMUNDA.toString();
   }
 }
