+++ webapps-schema/src/main/java/io/camunda/webapps/schema/descriptors/index/MappingIndex.java
@@ -5,29 +5,33 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.webapps.schema.descriptors.usermanagement.index;
+package io.camunda.webapps.schema.descriptors.index;
 
+import io.camunda.webapps.schema.descriptors.AbstractIndexDescriptor;
+import io.camunda.webapps.schema.descriptors.ComponentNames;
 import io.camunda.webapps.schema.descriptors.backup.Prio5Backup;
-import io.camunda.webapps.schema.descriptors.usermanagement.UserManagementIndexDescriptor;
 
-public class UserIndex extends UserManagementIndexDescriptor implements Prio5Backup {
-  public static final String INDEX_NAME = "user";
+public class MappingIndex extends AbstractIndexDescriptor implements Prio5Backup {
+  public static final String INDEX_NAME = "mapping";
   public static final String INDEX_VERSION = "8.8.0";
 
-  public static final String ID = "id";
-  public static final String KEY = "key";
+  public static final String MAPPING_KEY = "mappingKey";
+  public static final String CLAIM_NAME = "claimName";
+  public static final String CLAIM_VALUE = "claimValue";
   public static final String NAME = "name";
-  public static final String USERNAME = "username";
-  public static final String EMAIL = "email";
-  public static final String PASSWORD = "password";
 
-  public UserIndex(final String indexPrefix, final boolean isElasticsearch) {
+  public MappingIndex(final String indexPrefix, final boolean isElasticsearch) {
     super(indexPrefix, isElasticsearch);
   }
 
   @Override
   public String getVersion() {
     return INDEX_VERSION;
+  }
+
+  @Override
+  public String getComponentName() {
+    return ComponentNames.CAMUNDA.toString();
   }
 
   @Override
